+++ src/utils.cpp
@@ -1,7 +1,7 @@
-#include "Utils.h"
+#include "utils.h"
 
 #include <cstdlib>
 
-float get_chance() {
-    return static_cast<float>(rand()) / static_cast<float>(RAND_MAX);
+double get_chance() {
+    return static_cast<double>(rand()) / static_cast<double>(RAND_MAX);
 }
